whatsapp-agent/requirements.txt
pyautogui==0.9.54
pillow==10.1.0
python-dotenv==1.0.0
pytesseract==0.3.13
pywhatkit==5.4